+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/AbstractTigerRouteCallback.java
@@ -35,6 +35,12 @@ public abstract class AbstractTigerRouteCallback implements ExpectationForwardAn
     private final Map<String, ZonedDateTime> requestTimingMap = new HashMap<>();
 
     public void applyModifications(HttpRequest request) {
+        if (!tigerProxy.getModifications().isEmpty()) {
+            parseMessageAndApplyModifications(request);
+        }
+    }
+
+    public void parseMessageAndApplyModifications(HttpRequest request) {
         final RbelElement requestElement = tigerProxy.getRbelLogger().getRbelConverter().convertElement(
             tigerProxy.getMockServerToRbelConverter().requestToRbelMessage(request));
         final RbelElement modifiedRequest = tigerProxy.getRbelLogger().getRbelModifier()
@@ -79,6 +85,12 @@ public abstract class AbstractTigerRouteCallback implements ExpectationForwardAn
     }
 
     public void applyModifications(HttpResponse response) {
+        if (!tigerProxy.getModifications().isEmpty()) {
+            parseMessageAndApplyModifications(response);
+        }
+    }
+
+    public void parseMessageAndApplyModifications(HttpResponse response) {
         final RbelElement responseElement = tigerProxy.getRbelLogger().getRbelConverter().convertElement(
             tigerProxy.getMockServerToRbelConverter().responseToRbelMessage(response));
         final RbelElement modifiedResponse = tigerProxy.getRbelLogger().getRbelModifier()